Module autoware_auto::msgs::ffi

ROS msgs used in Autoware.Auto avp demo

Native ROS msg can be found on GitHub

Autoware.Auto msg can be found on GitLab

Structs

AutowareAutoMsgsComplex32

Complex32.msg

AutowareAutoMsgsHadmapRoute

A route within a high-definition map defined by the start and goal points and map primitives describing the route between the two.

AutowareAutoMsgsHadmapSegment

A segment of an HADMap which contains one or more MapPrimitives.

AutowareAutoMsgsMapPrimitive

Map primitive information

AutowareAutoMsgsRoutePoint

Representation of a position and heading

AutowareAutoMsgsTrajectory

A set of trajectory points for the controller

AutowareAutoMsgsTrajectoryPoint

Representation of a trajectory point for the controller

AutowareAutoMsgsVehicleControlCommand

Information that is sent to Vehicle interface

AutowareAutoMsgsVehicleKinematicState

AutowareAutoMsgsVehicleStateCommand

AutowareAutoMsgsVehicleStateReport

BuiltinInterfacesDuration

Duration defines a period between two time points. It is comprised of a seconds component and a nanoseconds component.

BuiltinInterfacesTime

Time indicates a specific point in time, relative to a clock’s 0 point.

GeometryMsgsPoint

This contains the position of a point in free space

GeometryMsgsPose

A representation of pose in free space, composed of position and orientation.

GeometryMsgsPoseStamped

A Pose with reference coordinate frame and timestamp

GeometryMsgsPoseWithCovariance

This represents a pose in free space with uncertainty.

GeometryMsgsPoseWithCovarianceStamped

This expresses an estimated pose with a reference coordinate frame and timestamp

GeometryMsgsQuaternion

This represents an orientation in free space in quaternion form.

GeometryMsgsTransform

This represents the transform between two coordinate frames in free space.

GeometryMsgsVector3

This represents a vector in free space.

StdMsgsHeader

Standard metadata for higher-level stamped data types. This is generally used to communicate timestamped data in a particular coordinate frame.